+++ Kursovaia_Gubaidullin/Collector.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Kursovaia_Gubaidullin
+{
+    class Collector
+    {
+        //собирает воедино код
+    }
+}
