+++ Kursovaia_Gubaidullin/Kursovaia_Gubaidullin.csproj
@@ -46,19 +46,22 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Collector.cs" />
     <Compile Include="Creator.cs">
       <SubType>Form</SubType>
     </Compile>
     <Compile Include="Creator.Designer.cs">
       <DependentUpon>Creator.cs</DependentUpon>
     </Compile>
+    <Compile Include="MemoryProvider.cs" />
+    <Compile Include="Model.cs" />
     <Compile Include="Questions.cs">
       <SubType>Form</SubType>
     </Compile>
     <Compile Include="Questions.Designer.cs">
       <DependentUpon>Questions.cs</DependentUpon>
     </Compile>
-    <Compile Include="Program.cs" />
+    <Compile Include="Controller.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Settings.cs">
       <SubType>Form</SubType>
@@ -66,6 +69,7 @@
     <Compile Include="Settings.Designer.cs">
       <DependentUpon>Settings.cs</DependentUpon>
     </Compile>
+    <Compile Include="UtilityProvider.cs" />
     <EmbeddedResource Include="Creator.resx">
       <DependentUpon>Creator.cs</DependentUpon>
     </EmbeddedResource>
