+++ Kursovaia_Gubaidullin/MemoryProvider.cs
@@ -1,0 +1,25 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Kursovaia_Gubaidullin
+{
+    public  class Memory
+    {
+        public static void Save(Model m)
+        {
+
+        }
+
+        public static Model Load()
+        {
+            Model m = new Model();
+
+
+
+            return m;
+        }
+    }
+}
